Tree.Folder: Add onLoad
Add onLoad as alternative to given query/variables.

Using the onLoad function allows to execute multiple queries / a query
document with more than one query, for example to fetch both items and
an item-index.

Adds tests for onLoad().

The changes on the old tests are minor timing differences in that our
code now uses multiple .then() and thus cannot be waited on with a
simple Promise.resolve().

Using a low 5ms delay to be on the safe side.