Implement GraphQL post processors
Add to methods registerGraphQLPostProcessor(type, fn) and
registerGenericGraphQLPostProcessor(type,fn) that allow
to define a post-processor function based on a type or a generic type.

Whenever the graphQL autoconversion encounters a type for which a
postprocessor is registered it will invoke the registered callback
with tree arguments. First the instance itself, then the query object
and then the parameters of the query. (value, query, params).

We apply the post-processors late after the result-tree has been
converted into an observable. This allows us to set members on
classes with @observable annotated props *without* making the new
property observable.

We use that to provide InteractiveQuery instances with the original
query.