Tree: Support custom rendered menus
Adds support for defining the label of a tree menu item as function in addtion to defining it as
string simple string.

The render function allows labels with arbitrary icons, number of icons and icon positioning and
whatnot.

Updates component docs including some missing doc updates from an earlier
commit.