Extend Graphql post-processors
For our current purposes the post-processor implementation works, but it
is kind of silly to be only able to define one post-processor per type.

Sure, we could work with composite post-processors, but we also don't
necessarily want central control, we might have several libs that add
post processors independent of the local project etc.

This adds an (empty, skipped) test as reminder to test this behavior
(and to test graphql in general)