Add equalsScalar
Adds a new method for the equality comparison of scalar values, which we
have been missing so far.

For most scalar types this is a simple === comparison, but the Date
based scalar types already need to compared the respective time values
via .getTime().

Other types might have other specific equality rules.

So we add an util model util/equalsScalar.js which has the equality
method as main method and also offers a named export

registerScalarEquals(scalarType, fn)

with which applications can define custom equality functions for their
custom scalar types.

This module fixes a reference error in the previous release