FKSelector: Add iQuery document support
Until now we only had the option to give FKSelector a GraphQLQuery
object which was then used to query data to drive the dialog selection
and filter queries based on textual user input.

This adds the ability to use an injected iQuery document as data basis.
This can be a document with all rows or just the first page. Whatever
the current state of the document is, FKSelector will present its
contents as selection. If the user changes the selection or is able to
chose another page, the query the injection was based on is re-executed
with new parameters.