FilterDSL: Allow implicit value types
When I created the value() function, my thinking was. "Well, this is a
scalar value that always needs a type, so we better make sure the user
won't forget and put it first".

But always having to specify the type is both ugly and cumbersome,
especially, when the type could be automatically detected in like 90%
of the cases.

So that's what this does now. We switch the order of arguments so now
it's

    value(value, [type], [name])

with the type being optional. If it is not given the scalar type will be
selected based on the javascript type.

    boolean => Boolean
    number => Int
    string => String
    instance of Date => Timestamp

These cover 90% of all use-cases with maybe the unfortunate detail that
if we want a "Date" we have to specify is explicitly as such.

Same goes for all custom scalar types and other rare/complex types.

This is a breaking change if you used the Filter-DSL. You have to at
least switch the order of the first two arguments of all value() calls.

In most cases you can leave out the second argument now.