Add imperative dialogs
Using reactstrap's <Modal> component works fine for the most part and
allows easy, a11y-compliant addition of modal dialog workflows into the
processes.

But sometimes you might have a dialog that you need to write by hand
(i.e. which is not a full subprocess in a dialog) but which, much like a
subprocess can be called from many different places. And sometimes, just
having imperative control over the dialog and promises is nice, too.

This adds a new API for those imperative handcrafted dialogs.

    openDialog(
        dialog => (
            <Container fluid={true}>
                <div className="row">
                    <div className="col">
                        CHOOSE!
                        <ButtonToolbar>
                            <button
                                type="button"
                                className="btn btn-secondary mr-1"
                                onClick={ () => dialog.cancel() }
                            >
                                Cancel
                            </button>
                            <button
                                type="button"
                                className="btn btn-secondary mr-1"
                                onClick={ () => dialog.confirm("A") }
                            >
                                Option A
                            </button>
                            <button
                                type="button"
                                className="btn btn-secondary mr-1"
                                onClick={ () => dialog.confirm("B") }
                            >
                                Option B
                            </button>
                        </ButtonToolbar>
                    </div>
                </div>
            </Container>

        ),
        {
            title: "Imperative Dialog Example"
        }
    ).then(
        result => result && alert( "You chose " + result)
    )

The first argument of openDialog is a render function that receives a
dialog API as argument. The api allows the dialog components to control
the outcome of the dialog

dialog.confirm(result)

Confirms and resolves the dialog with the given result object.

dialog.cancel()

Confirms and resolves the dialog with the configured resultOnCancel
(`null` by defaul)
value. This method will also be called when the user presses the escape
key or uses the close button on the header.

dialog.error(err)

Rejects the original promise with the given error. Should only be used
for actual error situations.

When the user calls any of these functions, the dialog closes and
removes itself automatically and the promise returned resolves or rejects
according to the dialog API method called.