Test-Setup: Configure MobX
Make sure to configure MobX to reject observable modfication outside of
actions.

We only need to care about the observables that are observed. When a
tree falls in an observable forest and no one is observing it, who
cares?

This allows creation and subsequent modification of observables
outside of actions.