Implement WorkingSet Merge
This commit really changes how the working set works. Instead of simply
persisting the collected changes within our process, we now do a full
entity merge including change conflict resolution.

Working Set Design
==================

In addition to keeping a map of changed entities we now keep a
non-observable copy of the state we first encountered the entity.

This state now can include an UUID version property ( by default
called "version").

The system keeps track of the changes on these versioned entities and by
default refuses to perform conflicting changes.

If a user changes Object X, but in between them loading the data and
them storing the data someone else changes fields in the same entity,
we don't want a last-one-wins result.

By default, the merge service will allow changes on the same entity but
on different fields. If changes conflict, the user will be shown a
change conflict resolution dialog where they can chose which of the
competing values they want to use.

The merge-process will deal with non-versioned types, too. For them,
last-one-wins is still true, but the merge process also only applies the
fields changed, so they enjoy still a more limited protection. Another
user can just overwrite changes, but not because they changed another
field and the entity happens to be stored as a whole.

System Merge Options
====================

The merge system comes with two types of configurations. There is the
server-side merge system configuration which is defined when the merge
service bean is defined.

de.quinscape.automaton.runtime.merge.MergeService#build provides a
builder API to configure the service. The configuration is mirrored on
the client side as `config.mergeOptions`

WorkingSet Merge Configuration
==============================

The other part of the merge configuration is local to the working set
and is defined when the working set is created.

It corresponds to the java type

    de.quinscape.automaton.model.merge.MergeConfig

It contains per-type configuration for field groups and ignored fields.

It defines whether the user is allowed to apply and to discard.

One the client-side, the configuration also can include callbacks to
customize merge behavior.