Implement conditionBuilder()
Implement an conditionBuilder() helper to conveniently create
conditional FilterDSL conditions, that is FilterDSL conditions whose
shape depends on known values at request time.

Usage
=====

The condition builder works very similar to the "classnames"
NPM module.

    conditionBuilder(
        flag &&
            field("aaa")
                .eq(
                    value("bbb")
            )
    )

The function combines all given parameters to a single FilterDSL
expression while ignoring all parameters that are not a FilterDSL
object. `null`, booleans, strings are all ignored. Only objects are
combined.

The result depends on the number of arguments that evaluate to such an
object.

If only one arguments is a condition, that condition is returned as-is.

If more than one argument are conditions, they are combined with an or()
expression.

If no arguments are conditions, the method returns `null` which is the
neutral condition / always true.

Adds a test for the buildeThe function combines all given parameters to
a single FilterDSL expression while ignoring all parameters that are not
a FilterDSL object. `null`, booleans, strings are all ignored. Only
objects are combined.

The result depends on the number of arguments that evaluate to such an
object.

If only one arguments is a condition, that condition is returned as-is.

If more than one argument are conditions, they are combined with an or()
expression.

If no arguments are conditions, the method returns `null` which is the
neutral condition / always true.

Adds a test for the builder