Enable wire format converter reg during startup
We have the problem that the actual wire-format is only created after
the startup init function is executed.

To be able to register converters within the startup init we start with
a minimally mocked version of wire-format that just remembers all
converters registered with it so that we can register them with the real
wire format once it is created.