Button: Refactor for revalidation
With revalidation we  might get new errors on form submit, so we move
our double-checking

If we have an action function we check before like we used to. If the
action needs to submit the form it also needs to check for new errors
afterwards.

In case of a transition button, we move the isDisabled() check after we
have submitted the form.

New errors will be picked up and the transition will not be invoked on
erroneous fields.