Improve FilterDSL, retire conditionBuilder
First of all, the conditionBuilder is an awkward builder helper based on
another builder helper, the FilterDSL. Then it turns out, we need more
than or() conditions, so all in all it seems better to redo the whole
thing.

We get rid of the conditionBuilder and implement its behavior for the
normal FilterDSL and() or() and not() methods. This makes the FilterDSL
itself more flexible and also makes building conditions with varying
shapes much easier.

So now, the and(), or() and not() function will ignore all non-node
arguments passed to them and optimize themselves according to the number
of remaining operands.

If and() and or() have no arguments, they evaluate to `null` i.e. that
is true.

not() without operand resolves to a "false" condition.

If and() and or() have only one operand, they reduce themselves to that
operand.

Composite conditions
====================

For composite conditions, this brings some problems for us as we've been
expecting the user to give us a

    and( component( "id" ...))

as default condition. The and() of it now gets optimized away, so this
commit also changes the updateComponentCondition() method to deal with
both deal with single component conditions and to update them as before.

Adds some additional tests for this.

Makes sure to also check for non-object conditions in the
InteractiveQuery.update() to deal with conditional conditions without
and()/or()/not().