Extend createDomainObject()
As a convenience function to not have to define defaults for all
properties that can't be `undefined`, we go and set all non-id fields of
a newly created domain object to `null` or to an empty array if it's a
list.

These act as convenient default values for form elements that should
interpret null to be an empty field. The new revalidate feature
makes sure the user actually enters meaningful values and fulfills all
not null contracts in the end.