Refactor process options
Merges the options in Process.js and config.

config.processDialogs contains the global defaults for
`process.dialogOptions` options which mostly control the process
dialog.

We add another option "title" which either contains a factory function
( name => title) or a simple title string. (Default is the old title as
function `name => i18n("Sub-Process {0}", name)`)

If the title is an empty string, the process dialog header is hidden.

Note that this might impact a11y in that without the explicit close
button in the modal header, there is no explicitly declared way of
closing the dialog. The users have to either click outside of the modal
or press the ESC key.

The per-invocation options can be passed as third parameter the
per-invocation options can be passed as third parameter to
process.runSubProcess

    process.runSubProcess(
        "my-subprocess",
        input,
        {
            title: "Custom Title"
        }
    )