createDomainObject: Add nullFields option
Make nulling all existing fields on new object optional. It is a very
useful behavior in apps and even a good default, but there are cases
where we deal with partial domain objects where it causes problems.

We are allowed to not select a non-null field and the object get
converted as such, but once we are nulling the field, the converter
enforces the non-null rule on the now existing property.

This adds the option and register it with registerDomainObjectFactory
with that option which fixes clone() behavior.

Adds a test.