WorkingSet: Make addChanges largely optional
The API was a bit too cumbersome and needlessly so in most cases.

When we get a snapshot for a base version in registerBaseVersion
we can just as well assume that the same observable will be used to
track changes. So we automatically call .addNew() on the given domain
object unless a new option "trackChanges" is set to false.

Note that you need to make sure that you call registerBaseVersion with
the clone you are editing if you are using cloning. So
registerBaseVersion must happen *after* config.inputSchema.clone()