Revert .back() decision
This commit is basically a revert of
20ed3098e8df1d80b5271e3c7c6c6ac1fa839c95.

As odd as the behavior using the browser history might look at times,
it's the only one that actually seems to work.

Recording the back-state results breaks multi-level back() constructs
because it inserts itself and then the next back() doesn't work right.

Not recording the transition does not update the browser state right.

So we return to using browser history navigation for our
transition.back().