Fix working set edge cases
Fixes some overlooked edge cases for working set change detection and
adds a new layer of tests for that.

Change detection pecularities
=============================

The way we detect changes may seem a bit odd but is the result
underlying problem domain and how we're dealing with that in other
places.

During the whole merge process, many-to-many relations are a bit of an
inconvenience. Semantically, we see them as abstract entity
associations/relations which we can edit with the
<AssociationSelector/>.

The underlying reality however is that we use a separate link table that
we elevated to being a full entity to deal with it like any other table.

Changes in associations however never create merge conflicts. Either two
objects are no connected and we create a new link table entry to connect
them or they are connected and we delete the link entry.

New objects never introduce conflicts and the conflict between a user
deleting a link and another not changing it is not detecting per se
either.

As a solution we introduced artificial changes to the entity on which
we use the <AssociationSelector/> on the GraphQL list field.

Here we now also special case changes to those many-to-many list fields.

Since the <AssociationSelector/> doesn't know about our working set and
it seemed too cumbersome to introduce it, and we're special-casing
many-to-many relations in other ways, anyway, we just deal with the
link changes from within the source entity, which means that this is the
one case, where we automatically detect new additions to and deletions
from the list of link objects.

For any other relation connection, be it forward or a non-many-to-many
backward relation (running into the opposite direction of its foreign
key), we still expect the user to call workingSet.addNew(domainObject).