ProcessHistory: Get rid of versionedPropsChanged
While it once used to be the only existing flag here, it is now somewhat
superfluous as it will always reflect a change in history position.

Instead of resetting and setting a status, the outside observer
remembers the current history position and then can track whether that
changed.

This way, resetting the position/undoing/redoing will still lead to the right
evaluation.