Fix subprocess history handling
With our rediscovery that true browser-history back is the one true
back, it wasn't a big surprise when subprocesses did not either.

This commit makes the sub-process functionality use the same back
functionality as transition.back() does.

backToHistoryId(id)
===================

This commit finally introduces a second back function that returns to
a state state identified by its id field

To get the current history id use process.getCurrentHistoryId();

The sub-process mechanism uses both to remember the state it was in
before the subprocess and to return to the exact same history entry
later.

Sub-Process Nuking
==================

For many if not most sub processes, the normal back() behavior to end it
is okay.

If the sub-process however represents something like a online shop
checkout, it is better to invalidate the sub process once it has
ended.

To enable this, pass the option nukeOnExit: true to runSubprocess.

    process.runSubProcess(
        "mySub",
        input,
        {
            nukeOnExit: true
        }
    )

The implementation goes back to the state before the current one and
then recreating the current state in the history, but on a different
history branch, removing the subprocess states from browser history as
well as memory references.