Fix changes on unselected fields
There was still a big bug regarding the editing/merging of partially
selected entities.

getFieldChanges() would iterate over all fields within the GraphQL type
and get "undefined" for both base and currValue.

However, for luxon types, the equality function "+a === +b" would then
turn twice do a coercion from `undefined` to `NaN` which as we know is
not equal to itself and thus a false entity field changes with undefined
values were created.

This commit fixes this generally by just ignoring all undefined values.

A field should either have a correct scalar value or `null`. If it is
undefined, we just assume it is a non-selected field and ignore it.

Do *NOT* overwrite scalar field values with `undefined`: Use `null`
instead.