Make DateTime equality more precise
undefined DateTime being not equal to themselfs is stupid.

And while it's kinda cute that `+null === +0`, I'd like us to
distinguish between a null Date and a Date with 0 microseconds since
epoch.

This commit does both by not coercing falsy values but substituting
null.

So a null Date is equal to an undefined Date in general.