WorkingSet: Add isModified method
The merge changes on the old working set broke the status display within
the IQueryGrid and within the <WorkingSetStatus/> component.

The existence of a MODIFIED entry no longer means that there were actual
modifications.

We have to basically do the whole process of finding field changes based
on grouping and scalar comparisons plus we have to regard the virtual
many to many changes.

This implements workingSet.isModified(entity) based on creating new
versions of the merge preparation methods.

hasFieldChanges() and hasManyToManyChanges() do what
getFieldChangesForObject() and getManyToManyChanges() does, but without
creating change objects and with early exit, and of course only for
a single entity at a time.

Adds the missing test-case for deleting a many-to-many relation.

Inserts checks for correct return of isModified into the existing
working set tests.