Add extended maxLength support
The new version of domainql-form offers a maxLength prop on <Field/> to
define the maximum length for String fields.

We extend this in automaton to not require explicit definition but to
get the maximum length limit from the fieldLengths the server side now
provides for us.

The props on <Field/> and <TextArea/> still supercede the values from
config.fieldLengths.

Use the props to override the domain defaults or to specify max lengths
for non-typed forms.