Implement InteractiveQueryEditor
So what we've done so far is editing InteractiveQuery directly in the
source code, aided by the "JS GraphQL" plugin, potentially substituing
Filter DSL expressions as JSON graphs etc pp.

This works for us, but is clearly nothing a normal application user could use.

GraphQL makes it relatively easy to provide a UI to configure the
available InteractiveQuery based graphQL queries.

This commit implements a basic editor for such queries with an embedded
condition editor and a test table output.

This is kind of preliminary in that we're close to an acceptable
user-experience, but not quite. The test ouput tables work but are very
funky.