DecimalField: Special-case Delete key
The way we're reformatting the string while the user is editing leads to
some unfortunate usability edge cases which we already mitigated for
deletion with the backspace key.

Deletion with the "Delete" key however still had some unfortunate
behaviors.

This commit adds special handling for the delete key when the current
cursor position is just before the "." or ",", in which case deleting
the next character will do nothing as reformatting undos the deletion.

What we can do however is move the cursor right so that the user is at
least in the right position as-if the deletion had happened.

The same goes when the user has deleted all non-fractional digits and
the current display is "0," something.

This way, continued pressing of "Delete" will lead to a much more
natural and predictable outcome.