Implement createFilteredMockQuery
Adds a more fancy type of mocked query to our testing tools.

the mocks so far have been working on a constant result set or on a
method to return the current result.

Now we add a mock factory that works based on a base iQuery document
that contains all available rows in the test.

The implementation will then both evaluate a query condition on those
rows if present and will also do pagination of the results, when
enabled. This is an easy way to test interactive widgets that execute
queries (like AssociationSelector)