AssociationSelector: Add onNew prop
Adds an <AssocationSelector onNew={ link => { ... } } /> prop that
allows modification of newly created link table instances. This is
needed for non-trivial link tables that contain more than the pure
linking information and primary key, but also things like owner or
organization or whatever.

As I started to write a test for this feature I came to realize that
there was no tests for association selector at all.

So I wrote a general test for the association selector first, with a
detour of having to implement JS object filtering based on JSON
condition graphs to return the correct values for one of the association
sector edge-cases.

When the user selects an entry on e.g. the second page of catalog values
and then goes to another page before closing the window, we have
refetch all newly selected elements that are not in the current list.

The widget does this by executing the same query with a IN condition
matching those ids. And now with js object filtering this also works
with a filtered mock query.

Finally, this adds the test for the onNew prop.