JS Filter: Optimize
Easy optimization opportunity: For every function that is invoked with
only non-function arguments we don't need to generate a wrapping
function since the result will always be the same.

While this might happen not as often in production as it does in the
test code, it still is an easy, no-cost optimization.