Export createMockedQuery / createFilteredMockQuery
Export our internal mocking functions for iQuery execution to allow use
in app testing or custom component testing.

createMockedQuery is kinda of legacy already and just keeps surviving
because some older, pretty complicated tests already exist and use stuff
like condition emulation (not evaluation like we do now) to work.

Rewriting those is too much of a hassle at this point.

The recommendation would be to try to use createFilteredMockQuery as
much as possible and only fallback to createMockedQuery if you have to.