Column: Add className prop
Adds support for a className prop that controls classes being renderd on
the table cells for that column.

If the given value is a string it applies to all columns.

The className prop can also be a function that returns the classes for a
cell given the row object ( row => classes ).