Fix history state handling
There were some mistakes in the state handling that resulted in invalid
indices being read from the processHistory array.

"currentHistoryPos" is now a simple virtual counter tracking the history
position to have the correct relative navigation.

"historyCounter" was separated from that and provides a unique new
navigationId for every entry and is never reset.