Fix updateComponentCondition
We recently changed the FilterDSL to do automatic optimization of and()
and or() expressions, mostly to allow for easy condition building

    and(
        myVar && ...
    )

where the shortcutted evalualtion passes in something falsy.

This in return necessitated refactoring in the updateComponentCondition
function which was missing one important test.

Adds a missing check to make sure that we add a new component condition
if we have not found it among one or more others within a wrapping and()

(which is the case of how the server generates these component
conditions and which in the end broke <Tree/>)

Adds a test for the missing case.