TMP: Tree: Update <Objects/> and <IndexedObjects>
Another product of our end of the year workshop, we implement a tree
refresh.

Changes the state handling inside <Tree/> from a reducer/useMemo based
state to a local mobx state that has partially @observable properties
(replacing the reducer state) and partially non-observable properties
which are either props we don't expect to change or are internal storage
to register our existing tree container components (<Objects> /
<IndexedObjects>) by their new id prop.

Both components now require a logical id which they use to register an
update function that can be called from the outside.

    const ctx = Tree.getContext("animal-tree");
    ctx.update("indexed-nodes");

The first id in the example is the id of the tree, the second id the
newly added id on one of the tree container components.

This code looks up the internal tree state / context via the static
Tree.Context method and then tells that state to trigger an update on
the container component with the given id.

If the id does not exist (yet), nothing will happen. We cannot
distinguish between the typo in an id and an id that doesn't exist yet
because the node wasn't opened yet.