IQueryGrid: Handle conditional columns
We're been using React.Children methods to iterate over the current
children of the <DataGrid/> component. These methods introduce `null`
elements for elements that are initially `false` like they should be for
a conditional element.

We change the internal handling to cope with this fact.

We use an initial Reacth.Children.forEach to filter the memoized columns.

We check if the columnElement is falsy and while we're at it, we
validate the component type, too and ignore all non-<Column/>s.

We remember the element in the columns array and then iterate over the
columns array to render all columns.

Note that we memoize the columns and only regard type changes or filter
changes as a reason to refresh it.

If you are out to create fully-dynamic columns  that are not
user-configured columns, you might need to provide the iQuery grid with
a key prop that changes when you need an update.