Button: Improve error checking
With the new FormContext changes we can now disambiguate between
"outside-of-form" or "no-root-object" and we use that knowledge to
improve our behavior.

Buttons outside of a Form are not connected to a Form object and thus
cannot have errors.

Non-discard buttons within a form are disabled if there is no form
object in that form.