FKSelector: Don't register field context
I was never quite sure if registering a field context for <FKSelector/>
is the right thing to do and now there are problems with doing so.

The FKSelector uses the error system to internally present an error if
we allow for textual input and there no unambiguous match, but otherwise
it is not a normal field in the sense that it takes part in normal
observable (re)validation etc.

The easiest way seems to just not register a field.