Fix .back() again
Fix more issues with the back() function. The code remains troublesome
and is not very well-tested, mostly because I don't really know how.

It hangs both on the "history" library and browser navigation and
history behavior, exactly what jsdom *doesn't* have.

We can't tie currentHistoryPos to the navigationId because we shrink
history above a certain limit which makes them go out of synch.

We just need to increase it when we push a new history entry.

Pruning also seemed to cut wrong, which together led to some very weird
behavior when using the back() function repeatedly.