Add extended history logging
I mainly used this to debug the issues myself, but I'm going to leave it
in for now and put it behind an option.

If you set `config.logHistory = true;` the process history system will
do an extended logging of current state and operations.

Use this to investigate/document issues in the future in case .back()
needs fixing again :\