Calendar: Fix required behavior
Revalidation can add errors relating to required calendar fields which
the user couldn't get rid of so far since we wrote the values back the
wrong way.

We had a dedicated action that changed the property directly in the
observable without going through the normal form validation.

Changing dates in the later added text input works fine.

So instead of writing back at DateTime instance into the observable, we
use the known format to convert it back to a string and then feed that
string through the normal handleChange method.