Refactor out in-memory query
We already had an implementation of the interactive query mechanism
based off a source iquery document containing all data.

We refactor this method out to make it accessible for other purposes,
and fix a few edge cases while we're at it.

The new method is

    import { evaluateMemoryQuery } from "@quinscape/automaton-js"

    // ...

    const doc = evaluateMemoryQuery(
        wireFormat,
        sourceIQueryDoc,
        queryConfig
    )

By default, the method returns a new InteractiveQuery instance, but you
can also pass one in as fourth argument to be updated instead.