FilterDSL: Add field.toString()
Contrary to my assumptions, using something like containsIgnoreCase on a
numeric field doesn't just work but in fact causes an SQL error :\

I don't want to implement a general cast at this point given the
complication on one side only knowing GraphQL types and one needing Java
equivalents and how to encode those etc.

So I'm going just for a specialized cast to string which we will call
"toString"

    field("num")
        .toString()
        .containsIgnoreCase(
            value(v)
        )

This contains the changes on the FilterDSL itself as well as the js
filter transformer.

See commit 0bf43887 in the automaton repository for the corresponding
server-side commit.