pickSchemaTypes: Add selective queries / mutations
The schemas contained no QueryType/MutationType handling until now.

This adds a way to selectively add query and mutations methods to your
schema.

    pickSchemaTypes(rawSchema, [ "q:iQueryFoo", "m:updateFoo" ])

Here we pick only a query named "iQueryFoo" and a mutation named
"updateFoo" and all their dependencies (return types and argument
types). you can use "q:" as a prefix for mutations and "m:" as a prefix for
mutations mixed with the normal type references.