CachedQuery: Refactor out new GraphQLQuery mocks
While we had js-simulated queries before, those mocks where not as
formally complete as the ones we're effectively using with the
CachedQuery mechanism.

Some of the early widget code was very patchy in how it dealt with the
type knowledge. Since the early days we added some more extras to our
internal schema format that complete gaps in the type definitions.

New widgets and the rewrite of <FKSelector/> now can use that data to
require much less configuration etc pp.

As a consequence however, they are becoming more demanding in what they
require from the the query objects they deal with.

For a complete mock query definition we need the actual query string as
parsing it allows us access to the method names and types etc involved.

This is where our old mocks fail and why they should be deprecated by
for new code. (I'll leave the functioning tests as they are for now)

The way CachedQuery defines its query based on an original iQuery
document linked to a query allows its queries to copy those parts over
from the source query.

Which is why we are refactoring out the internal code so that other code
/ test code might gain from it.

    createMemoryQuery(source, queryConfig)

creates a new GraphQLQuery mock based on given source InteractiveQuery
instance and an overriding query config.

This also adds an even more convenient version for tests

    loadMemoryQuery(
        "InteractiveQueryFoo",
        rawData,
        query // original query
        queryConfig // overriding query config
    )

This creates a new GraphQL query mock based on raw JSON data
as it comes from the server.

It will automatically handle all conversion and will connect it to the
given GraphQLQuery instance etc pp.