Redo <FKSelector/>
The fk selector has been one of our first complex widgets and the code
showed it. It didn't quite make use of the later added type
introspection abilities and required too much configuration.

In addition to that it showed several bugs not easily fixable with the
old code.

This commit refactors, simplifies and modernizes the <FKSelector/> code,
and also streamlines its usability and harmonizes the feature set.

Reworks the existing test and adds tests for new features.