findComponentNode: Handle single component
Since our adaption to the auto-optimizing filter-dsl we had to become
more diverse in how our cooperative conditions are kept.

findComponent now needs to deal with either composed component
expressions, a singular component expression or a non-component
expression with implicit componentId equal to null.

We refactor out a check we already used for pretty much the same purpose
when updating cooperative conditions and use it to check.