Make collapsible autoresize on scroll
Make the collapsible fit to parent and viewport on scroll. If the
collapsible gets pushed out of the viewport, due to header or footer
pushing the sticky context, the size will be automatically adjusted to
fit into the viewport again. It will also take additional optional
padding parameters towards the viewport top and bottom.
This adjustment will also take place, if the window size changes.