Button: Follow changes in domainql-form
We now have to account for slight changes in how the formConfig is set
up and slight differences in how the button reacts.

domainql-form v3.5.0 largely removes the concept of buttons being
"outside" of the form. With the new model of many forms in many parts it
does not really make that much sense anymore.

The button is always connected to a form context it gets via the
formConfig. This is either the default form context or the form context
explicitly given to either <Form/> or <FormConfigProvider/>.

When the button has a transition, it follows the transition's discard
flag. If discard is not set, the button will disable if there are
errors anywhere in the form context.

We also keep disabling buttons if they are non-discard and inside a
<Form/> that has a null value for easy auto-disabling for forms.

To integrate action buttons, we define a new convienence method

    <Button
        action={ () => ... }
        disabled={ Button.disabledIfErrors() }
    >
        ...
    </Button>

This provides the same disabling behavior for action buttons.

A button that is inside a form will continue to submit the form before
executing the transition.

If it is outside of a form, it will just trigger a revalidate() on
its form context.